+++ src/Server/static/js/ui/modals.js
@@ -321,5 +321,3 @@
         }
     };
 })();
-
-console.log('UI modals loaded');